+++ djc/recipe/tests/test_docs.py
@@ -50,7 +50,7 @@ def setUp(test):
 def test_suite():
     suite = unittest.TestSuite((
             doctest.DocFileSuite(
-                '../README.txt',
+                '../README.rst',
                 setUp=setUp,
                 tearDown=zc.buildout.testing.buildoutTearDown,
                 optionflags=optionflags,